requirements.txt
Package                      Version     Editable project location
---------------------------- ----------- ----------------------------------------
absl-py                      1.0.0
addict                       2.4.0
albumentations               1.1.0
antlr4-python3-runtime       4.9.3
astunparse                   1.6.3
attrs                        21.4.0
backcall                     0.2.0
cachetools                   5.1.0
certifi                      2022.5.18.1
charset-normalizer           2.0.12
cupy                         8.3.0
cycler                       0.11.0
debugpy                      1.5.1
decorator                    5.1.1
dominate                     2.6.0
einops                       0.4.1
entrypoints                  0.4
fastrlock                    0.8
flatbuffers                  1.12
fonttools                    4.33.2
gast                         0.4.0
google-auth                  2.6.6
google-auth-oauthlib         0.4.6
google-pasta                 0.2.0
grpcio                       1.46.3
h5py                         3.7.0
idna                         3.3
imageio                      2.18.0
importlib-metadata           4.11.4
ipykernel                    6.9.1
ipython                      7.32.0
jedi                         0.18.1
jittor                       1.3.4.14
joblib                       1.1.0
jupyter-client               7.2.2
jupyter-core                 4.10.0
keras                        2.9.0
Keras-Preprocessing          1.1.2
kiwisolver                   1.4.2
libclang                     14.0.1
Markdown                     3.3.7
matplotlib                   3.5.1
matplotlib-inline            0.1.3
mkl-fft                      1.3.1
mkl-random                   1.2.2
mkl-service                  2.4.0
nest-asyncio                 1.5.5
networkx                     2.6.3
numpy                        1.21.6
oauthlib                     3.2.0
omegaconf                    2.2.1
opencv-python                4.5.5.64
opencv-python-headless       4.5.5.64
opt-einsum                   3.3.0
packaging                    21.3
pandas                       1.3.5
parso                        0.8.3
pexpect                      4.8.0
pickleshare                  0.7.5
Pillow                       9.1.0
pip                          22.1.2
prompt-toolkit               3.0.29
protobuf                     3.19.4
ptyprocess                   0.7.0
pyasn1                       0.4.8
pyasn1-modules               0.2.8
Pygments                     2.12.0
pyparsing                    3.0.8
python-dateutil              2.8.2
pytz                         2022.1
PyWavelets                   1.3.0
PyYAML                       6.0
pyzmq                        22.3.0
qudida                       0.0.4
requests                     2.27.1
requests-oauthlib            1.3.1
rsa                          4.8
scikit-image                 0.19.2
scikit-learn                 1.0.2
scipy                        1.7.3
setuptools                   61.2.0
six                          1.16.0
tensorboard                  2.9.0
tensorboard-data-server      0.6.1
tensorboard-plugin-wit       1.8.1
tensorboardX                 2.5
tensorflow                   2.9.1
tensorflow-estimator         2.9.0
tensorflow-io-gcs-filesystem 0.26.0
termcolor                    1.1.0
terminaltables               3.1.10
threadpoolctl                3.1.0
tifffile                     2021.11.2
timm                         0.3.2
torch                        1.7.1
torchvision                  0.8.2
tornado                      6.1
tqdm                         4.64.0
traitlets                    5.1.1
typing_extensions            4.2.0
urllib3                      1.26.9
wcwidth                      0.2.5
Werkzeug                     2.1.2
wheel                        0.37.1
wrapt                        1.14.1
yapf                         0.32.0
zipp                         3.8.0
